Split js sourcemapping into base & js classes
